Create FinishedProjectClean
This is the cleaned up code after removing my comments for my ATM project. I currently want to learn how to update and save the account balance so that it has full functionality of an ATM. I thought that creating an empty list and appending the balance would work but I could not figure out how to make it function properly. If someone with more experience could show me how to do that I would really appreciate it. Thank you!